Add ARM template for web app deployment
Added a new Azure Resource Manager (ARM) template to `deploy.json` to automate the deployment of a web application and its app service plan. The template includes a `$schema` property, a `contentVersion` set to "1.0.0.0", and defines three parameters (`webAppName`, `appServicePlanName`, `skuName` with default "F1"). It also details the resources to be deployed, including a `Microsoft.Web/sites` resource for the web app and a `Microsoft.Web/serverfarms` resource for the app service plan, specifying names, locations, and SKU.